Move interfaces and models from StockImage to StockAsset
