+++ AdobeStockAsset/Model/Keyword.php
@@ -7,7 +7,7 @@
 
 declare(strict_types=1);
 
-namespace Magento\AdobeStockImage\Model;
+namespace Magento\AdobeStockAsset\Model;
 
 use Magento\AdobeStockImage\Api\Data\KeywordInterface;
 use Magento\AdobeStockImage\Model\ResourceModel\Keyword as ResourceModel;
