+++ AdobeStockAsset/Model/ResourceModel/Asset.php
@@ -5,14 +5,16 @@
  * See COPYING.txt for license details.
  */
 
-namespace Magento\AdobeStockImage\Model\ResourceModel;
+declare(strict_types=1);
+
+namespace Magento\AdobeStockAsset\Model\ResourceModel;
 
 use Magento\Framework\Model\ResourceModel\Db\AbstractDb;
 
-class Keyword extends AbstractDb
+class Asset extends AbstractDb
 {
     protected function _construct()
     {
-        $this->_init('adobe_stock_keyword', 'id');
+        $this->_init('adobe_stock_asset', 'id');
     }
 }
