+++ AdobeStockAsset/etc/di.xml
@@ -8,5 +8,8 @@
 <config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:ObjectManager/etc/config.xsd">
     <preference for="Magento\AdobeStockAssetApi\Api\Data\ConfigInterface" type="Magento\AdobeStockAsset\Model\Config"/>
     <preference for="Magento\AdobeStockAssetApi\Api\ClientInterface" type="Magento\AdobeStockAsset\Model\Client"/>
+    <preference for="Magento\AdobeStockAsset\Api\AssetRepositoryInterface" type="Magento\AdobeStockAsset\Model\AssetRepository"/>
+    <preference for="Magento\AdobeStockAsset\Api\Data\AssetInterface" type="Magento\AdobeStockAsset\Model\Asset"/>
+    <preference for="Magento\AdobeStockAsset\Api\Data\KeywordInterface" type="Magento\AdobeStockAsset\Model\Keyword"/>
 </config>
