+++ AdobeStockImage/etc/di.xml
@@ -11,5 +11,4 @@
     <type name="Magento\Catalog\Model\Product\Gallery\Processor">
         <plugin name="adobe_stock_image_remove_metadata" type="Magento\AdobeStockImage\Plugin\Product\Gallery\Processor" sortOrder="10" disabled="false" />
     </type>
-    <preference for="Magento\AdobeStockImage\Api\AssetRepositoryInterface" type="Magento\AdobeStockImage\Model\AssetRepository"/>
 </config>
